extract publisher interface
