+++ naturewatch_camera_server/__init__.py
@@ -8,6 +8,7 @@ from logging.handlers import RotatingFileHandler
 from naturewatch_camera_server.CameraController import CameraController
 from naturewatch_camera_server.ChangeDetector import ChangeDetector
 from naturewatch_camera_server.FileSaver import FileSaver
+from naturewatch_camera_server.TelegramPublisher import TelegramPublisher
 from flask import Flask
 from naturewatch_camera_server.api import api
 from naturewatch_camera_server.data import data
@@ -76,9 +77,15 @@ def create_app():
         flask_app.logger.warning("Videos directory does not exist, creating path")
 
     # Instantiate classes
-    flask_app.camera_controller = CameraController(flask_app.logger, flask_app.user_config)
     flask_app.logger.debug("Instantiating classes ...")
-    flask_app.change_detector = ChangeDetector(flask_app.camera_controller, flask_app.user_config, flask_app.logger)
+    flask_app.camera_controller = CameraController(flask_app.logger, flask_app.user_config)
+    flask_app.publisher = TelegramPublisher(flask_app.user_config, flask_app.logger)
+    flask_app.change_detector = ChangeDetector(
+        flask_app.camera_controller, 
+        flask_app.publisher,
+        flask_app.user_config, 
+        flask_app.logger
+    )
     flask_app.file_saver = FileSaver(flask_app.user_config, flask_app.logger)
 
     flask_app.logger.debug("Initialisation finished")
